+++ tweetbrain/backend/app/api_v1/songs/routes.py
@@ -1,16 +1,15 @@
-from fastapi import Depends, APIRouter
+from fastapi import Depends, APIRouter,Request
 
 from . import utils
 
 
 
-
-
-
+router = APIRouter()
 
 
 @router.get("/match")
 async def match_song(
-    request: Request, twitter_handle: str
+    request: Request, handle: str
 ):
-    result = utils.match_with_top_songs(twitter_handle)
+    result = utils.match_with_top_words(handle)
+    return result
