+++ tweetbrain/backend/app/genius/tests/test_search.py
@@ -4,6 +4,6 @@ from ..match import Match
 matcher = Match()
 
 def test_basic():
-    result = matcher.search_song("burger town")
+    result = matcher.search_songs("burger town")
     for hit in result["hits"]:
         print(hit["result"], "\n----\n\n")
